# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/admins": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: SuperAdmin"
                ],
                "summary": "Get all administrators",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetAll200"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminLogin401"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminLogin403"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: SuperAdmin"
                ],
                "summary": "Create a new administrator",
                "parameters": [
                    {
                        "description": "JSON request body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.adminCreateAdminBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminCreateAdmin201"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminCreateAdmin409"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminCreateAdmin422"
                        }
                    }
                }
            }
        },
        "/admins/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: SuperAdmin"
                ],
                "summary": "Get administrator by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Admin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetByID200"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetByID404"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetByID422"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: SuperAdmin"
                ],
                "summary": "Edit an administrator",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Admin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "JSON request body\u003cbr\u003e\u003cbr\u003e* All fields are optional.\u003cbr\u003e** If field left blank it will not be updated.",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditAdminBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditAdmin200"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditAdmin404"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditAdmin409"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditAdmin422"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: SuperAdmin"
                ],
                "summary": "\"Soft\" delete an administrator",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Admin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminDeleteAdmin200"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminDeleteAdmin404"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminDeleteAdmin422"
                        }
                    }
                }
            }
        },
...
        "/clients/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get client by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Admin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetClientByID200"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetClientByID404"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminGetClientByID422"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Edit a client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "JSON request body\u003cbr\u003e\u003cbr\u003e* All fields are optional.\u003cbr\u003e** If field left blank it will not be updated.",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditClientBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditClient200"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditClient404"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditClient409"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminEditClient422"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: SuperAdmin"
                ],
                "summary": "\"Soft\" delete a client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminDeleteClient200"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminDeleteClient404"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/swagger.adminDeleteClient422"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "swagger.adminCreateAdmin201": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.adminCreateAdmin1"
                },
                "message": {
                    "type": "string",
                    "example": "Created new admin"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminCreateAdmin409": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Admin with email already exists"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminCreateAdmin422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "First name is required"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminCreateAdminBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
...
        "swagger.adminDeleteAdmin200": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Deleted admin with ID of 13"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminDeleteAdmin404": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Active admin not found"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminDeleteAdmin422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ID in invalid format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminDeleteClient200": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Deleted client with ID of 13"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminDeleteClient404": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Active client not found"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminDeleteClient422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ID in invalid format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditAdmin200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.adminCreateAdmin1"
                },
                "message": {
                    "type": "string",
                    "example": "Updated admin"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminEditAdmin404": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Active admin not found"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditAdmin409": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email address already in use"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditAdmin422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email is in an invalid format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditAdminBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "swagger.adminEditClient200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.adminCreateClient1"
                },
                "message": {
                    "type": "string",
                    "example": "Updated client"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminEditClient404": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Active client not found"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditClient409": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email address already in use"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditClient422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email is in an invalid format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminEditClientBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "notes": {
                    "type": "string",
                    "example": "This is a test note full of random test that is related to this client."
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "phone": {
                    "type": "string",
                    "example": "(702) 867-5309"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
...
        "swagger.adminGetAll1": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2020-08-07T20:12:54Z"
                },
                "email": {
                    "type": "string",
                    "example": "admin@test.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "integer",
                    "example": 5
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2020-08-07T20:12:54Z"
                }
            }
        },
        "swagger.adminGetAll200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swagger.adminGetAll1"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "All admins"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "total_results": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "swagger.adminGetByID200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.adminGetAll1"
                },
                "message": {
                    "type": "string",
                    "example": "Returned admin"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminGetByID404": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Admin not found"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminGetByID422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ID in invalid format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminGetClientByID200": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/swagger.adminGetAllClients1"
                },
                "message": {
                    "type": "string",
                    "example": "Returned client"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.adminGetClientByID404": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Client not found"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminGetClientByID422": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ID in invalid format"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "swagger.adminLogin401": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Please provide valid credentials"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "swagger.adminLogin403": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Account has been suspended"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}